Add class determining how cells/faces relate to a level set function
Add a class NonMatching::MeshClassifier that determines how the sign of
a level set function vary over the active cells/faces of a
triangulation. Each cell/face are associated with one of values of enum
LocationToLevelSet. If the level set is positive/negative over the
cell/face it is associated with LocationToLevelSet::inside/outside. If
the zero contour intersects the cell/face it is associated with
LocationToLevelSet::intersected.